backend/requirements.txt
dotenv
fastapi
uvicorn
langchain
langgraph
langchain_google_genai
google-cloud-storage
